update(audio, help): added audio, update help, add audio help

+++ commands/help.js
@@ -6,21 +6,32 @@ module.exports = {
     async execute(message) {
         
 
-        const embed1 = new Discord.MessageEmbed()
+        const embed = new Discord.MessageEmbed()
             .setTitle('Liste des commandes :')
-            .setDescription(`##help \n ##audio [nounours, pillarmen] \n ##play [lien youtube]\n ##prune [0-99] \n ##server \n ##avatar [empty-@tag]`)
+            .setDescription(
+                `##help\n`+
+                `##audio [nom d'audio] (##audio help pour plus de détails)\n`+
+                `##play [lien youtube]\n`+
+                `##prune [0-99]\n`+
+                `##server\n`+
+                `##avatar [empty || @tag]`)
             .setColor('#FF2D00')
             .setFooter('(Appuyez sur la croix pour supprimer le message)')
 
-
-        let msg = await message.channel.send(embed1).react("❌")
+        let msg = await message.channel.send(embed)
+        await msg.react("❌")
 
         msg.awaitReactions((reaction, user) => user.id == message.author.id && (reaction.emoji.name == '❌'), {
                     max: 1,
                     time: 30000
                 })
             .then(
-                message.channel.bulkDelete(2)
+                collected => {
+                    if (collected.first().emoji.name == '❌') {
+                        // console.log('suppression !')
+                        message.channel.bulkDelete(2)
+                    }
+                }
             )
     }
 }
